work with ints instead of floats

+++ a_matrix.cpp
@@ -1,11 +1,11 @@
 #include "a_matrix.h"
 #include <algorithm>
 
-float a_matrix::gap_cost(int k) {
+int a_matrix::gap_cost(int k) {
 	return this->gap_open + this->gap_extend * k;
 }
 
-float a_matrix::get_initial_score(int row, int col) {
+int a_matrix::get_initial_score(int row, int col) {
 	if(row == 0 && col == 0)
 		return 0;
 	if(row == 0 && col  > 0) {
@@ -30,10 +30,10 @@ cell* a_matrix::get_initial_pointer(int row, int col) {
 
 void a_matrix::fill_in_cell(cell* current, cell* above, cell* left,
 	cell* above_left) {
-	float d_score = d_matrix->cells[current->row][current->col].score;
-	float i_score = i_matrix->cells[current->row][current->col].score;
-	float substitution_score = substitution->get_score(sequence1.at(current->row - 1), sequence2.at(current->col - 1));
-	float a_score = above_left->score + substitution_score;
+	int d_score = d_matrix->cells[current->row][current->col].score;
+	int i_score = i_matrix->cells[current->row][current->col].score;
+	int substitution_score = substitution->get_score(sequence1.at(current->row - 1), sequence2.at(current->col - 1));
+	int a_score = above_left->score + substitution_score;
 
 	if (d_score >= i_score) {
 		if (a_score >= d_score) {
