+++ a_matrix.h
@@ -7,7 +7,7 @@ class a_matrix : public matrix
 {
 public:
 	a_matrix(unsigned int size, substitution_matrix* substitution,
-		float gap_open, float gap_extend, matrix* d_matrix, matrix* i_matrix) 
+		int gap_open, int gap_extend, matrix* d_matrix, matrix* i_matrix) 
 		: matrix(size, substitution, gap_open, gap_extend), d_matrix(d_matrix), i_matrix(i_matrix) {};
 
 	virtual std::pair<std::string, std::string> get_traceback();
@@ -16,9 +16,9 @@ public:
 	friend class i_matrix;
 
 protected:
-	float gap_cost(int k);
+	int gap_cost(int k);
 
-	virtual float get_initial_score(int row, int col);
+	virtual int get_initial_score(int row, int col);
 	virtual cell* get_initial_pointer(int row, int col);
 	virtual void fill_in_cell(cell* current, cell* above, cell* left, cell* above_left);
