+++ a_matrix_freeshift.h
@@ -4,12 +4,12 @@
 class a_matrix_freeshift : public a_matrix
 {
 protected:
-	virtual float get_initial_score(int row, int col);
+	virtual int get_initial_score(int row, int col);
 
 	virtual void fill_in_cell(cell* current, cell* above, cell* left, cell* above_left);
 	virtual cell* get_traceback_start();
 
 	cell* max_freeshift;
-	float max_freeshift_score;
+	int max_freeshift_score;
 };
